Fix update_binary function to be able to write more than 255 bytes
The T0 protocol doesn't support extended APDU (where we can write up to 65535 bytes by specifying 3 bytes for Lc), only T1.
But in transport/pcsc.py, it explicitly selects T0 as the protocol.
So for large data, we need to transmit in 255 byte chunks.
The read_binary function has code to read more than 255 bytes, so we just adapt it for writing.